[TAA] Fixed a silly mistake where the negative offsets of the history clipping neighborhood where uint2. [MipGeneration] Fixed an issue where upon loading a world, there would be some duplicate textures. [Misc] Fixed warnings. [Scripts] Added solution generation scripts for VS2022 (VS2019 will be removed once VS2022 goes out of preview version).

+++ Editor/ImGui_Extension.h
@@ -310,7 +310,7 @@ namespace ImGuiEx
 
     struct DragDropPayload
     {
-        typedef std::variant<const char*, unsigned int> dataVariant;
+        typedef std::variant<const char*, uint64_t> dataVariant;
         DragDropPayload(const DragPayloadType type = DragPayload_Unknown, const dataVariant data = nullptr)
         {
             this->type = type;
