+++ Runtime/World/Components/RigidBody.h
@@ -46,7 +46,7 @@ namespace Spartan
     class SPARTAN_CLASS RigidBody : public IComponent
     {
     public:
-        RigidBody(Context* context, Entity* entity, uint32_t id = 0);
+        RigidBody(Context* context, Entity* entity, uint64_t id = 0);
         ~RigidBody();
 
         //= ICOMPONENT ===============================
